Add a top-level README.md as a guide for newcomers
The intention is to provide a friendlier experience to new users along
with the over-arching goals of the project and a rough overview of the
structure of the repo.

This is a best effort from my understanding of the repository so far,
please let me know if you would like to see anything
added/changed/removed!