cubeb-backend: Pass `log_callback` to `cubeb_log_internal_buf_fmt`.
This moves the temporary stack allocation for the buffer out of the
macro expansion so we don't need to rely on the optimizer deduplicating
it when multiple cubeb_log!() expansions are present in the same function.

Also tweak the declaration of `cubeb_log_async` to match the signature
of `g_cubeb_log_callback`.