Fix build on arm64 linux.
On arm64 linux, c_char is u8. CStr::from_ptr is documented to take an
i8, but this is not accurate (see
https://github.com/rust-lang/rust/issues/113735); instead, it takes
c_char, which is sometimes but not always the same. One time it is
different is on aarch64 Linux, so without this commit, cubeb-rs fails to
build there.